Fix responsive bug
The post-widget areas were less than 100% width together on larger
screens so it allowed the footer to appear beside it due to their
float-left property.

This wraps them in a generic flexbox row instead of using float. This
also fixes their alignment to be in the center on larger screens.