Add syntax highlighting to markdown files.
All of our codeblocks inside of our blog are marked up using the
standard github flavored markdown. For example:

```javascript
let [ first, ...rest ] = items;
```

But our website markdown renderer is not marking up these codeblocks
appropriately when it gets transformed.

This installs  the [gatsby-remark-prismjs][1] plugin which is what Gatsby
itself uses internally for its documentation. I figured that why not
use what the Gatsby team uses since that ought to have the least
friction.

I found a nice trove of quality themes in the
[syntaxt-highlighting][1] package on npm (and [explainer][3]) that
works with many different syntax highlighters in addition to prism,
and set us up to use one of the themes that was simplistic, and didn't
clash with our logo scheme. I'm not a color-matching expert, so I'm
very open to alternatives :)

Learning
===

The plugin has support for inline code blocks so that you can use
`›` as a language specifier. So, for example, you can highlight with

`js›let [first, ...rest] = iterable;`

In adittion to the various themes in the syntax-highlighting package,
we also havethe [default prism themes][3]. More prism thems can be
accessed via the [prism-themes][4] package.

Alternatives
===

If we wanted to go with the highlight.js library, we can use the
[gatsby-remark-highlights][5] plugin.

[1]: https://www.gatsbyjs.org/packages/gatsby-remark-prismjs/
[2]: https://www.npmjs.com/package/syntax-highlighting
[3]: https://atelierbram.github.io/syntax-highlighting/prism/
[4]: https://www.npmjs.com/package/prism-themes
[5]: https://www.gatsbyjs.org/packages/gatsby-remark-highlights/