Add hotjar
Hotjar is nice when trying to understand our website.

Even though it's pretty paper-thin, I went with the existing [hotjar
gatsby plugin][1].

Note that since the HOTJAR_ID and HOTJAR_VERSION are publicly
available (they're embedded into the JS snippet), I didn't bother
hiding them behind environment variables.

Also, the plugin behavior is that this only adds hotjar in
production. Ideally, we'd use a different hotjar configuration for
staging than for production, but I think this is fine for now.

[1]: https://www.gatsbyjs.org/packages/gatsby-plugin-hotjar/?=hotj