use `devDependencies` for interactors in tutorial
the `dependencies` of a plugin in `package.json` will be installed into a project whenever the plugin is installed. However, JavaScript used to _test_ the plugin is not necessary for consumers of that plugin. If you use a plugin that is tested with interactors, we don't want to install interactors and all interactors dependencies into the project.

This tweaks the tutorial slightly to add the `yarn --dev`  flag to the commands to install `@interactors/material-ui` and `@interactors/with-cypress` That way, they will be added to `devDependencies` instead and will not be installed into the consumers of a plugin.