🐛Fix bogus toggling of the OSS projects
The glass pane that closed the toggle was covering the entire site,
and always on. This means that nobody could sign up because the button
was covered by the select box closer 😬

This makes sure that the closer is only active when the context menu
is _showing_, and never at any other time.