Flip the script and put Deno site in front of Gatsby
We want flexibility with how we manage our public
website, in order to

1. Not lock us into a particular technology
2. Be able to do dynamic stuff as we see fit.
3. migrate incrementally instead of with a giant bang.

This replaces the Gatsby site with a simple HTML server which is the
one that was driving our web page. Of course, it delegates most of the
work (including the home page) to the Gatsby site, but the important
thing is that the dynamic component is sitting at the top.

We can still edit the gatsby site, but we can also add pages using
HTML templates, or delegating to some other technology.