Proxy the Effection and Graphgen sites directly
Cut Netlify out of the loop.

Right now, when someone asks for https://frontside.com/effection

The request proceeds from

1. frontside.deno.com
2. frontside.netlify.app
3. effection.deno.com

And the response comes back via

3. effection.deno.com
2. frontside.netlify.app
1. frontside.deno.com

Eventually we want to be able to serve all the code directly from
frontside.deno.com, and the first stem is to cut out netlify
completely from that round trip.

This removes the netilfy redirects for graphgen, and effection, and
adds them directly to the frontside.com site.