frontside.com joints the revolution
Upgrade to the middleware framework we've been using for all our sites
which has now been extracted into a separate library. This
consolidates a lot of code, makes twind more accessible and standard
via `twing.config.ts` and also  improves the caching and proxy
behavior to load sites faster.

For example, we were not properly forwarding `if-none-match` headers
and so none of our proxy responses were being cached.