✨ Proxy the Interactors website
The Interactors website had been proxied by netlify, but we want to
take responsibility for serving it ourselves.

This enhances the proxy to take on serving static docusaurus sites.
Basically, it does two things, it accepts a `root` parameter which
gives the root path on the proxy site that we start serving from.

Also we enhance the proxy generally so that if it receives a redirect
to itelf, then it will also proxy the redirect rather than go to the
source site.