✨ Handle all redirect codes for sub sites
When we are proxying a website, and that site sends a redirect, we
need to redirect through ourselves. This was only happening on a 301,
but there are actually all kinds of redirect codes that have to be
handled. As it turns out, fresh framework which Graphgen was using
uses 302 for its redirects.

This goes ahead and adds 302,307,308 to the list, so that we pick up
all current and future redirects.